Fix zeitwerk error when eager_load is turned on * Rename /assistants directory to /assistant * Move Langchain::Messages namespace to Langchain::Assistant::Messages * Clean up zeitwerk configuration * Add code comments
